Refactor main into cli/persistence/output ownership seams with compatibility wrappers, direct seam tests, and runtime/docs updates
