+++ src/main.py
@@ -6,25 +6,47 @@ Updated for Gemini 3 (Nov 2025).
 
 import argparse
 import asyncio
-import json
 import logging
 import os
 import socket
 import sys
 import uuid
 from concurrent.futures import ThreadPoolExecutor
-from concurrent.futures import TimeoutError as FuturesTimeoutError
 from contextlib import nullcontext
-from dataclasses import dataclass
 from datetime import datetime
 from pathlib import Path
 from typing import Any, Literal
 
 import structlog
-from rich import box
 from rich.console import Console
-from rich.panel import Panel
-from rich.table import Table
+
+from src.cli import (
+    OutputTargets,
+)
+from src.cli import (
+    _cli_logging_mode as _cli_logging_mode_impl,
+)
+from src.cli import (
+    _resolve_output_targets as _resolve_output_targets_impl,
+)
+from src.cli import (
+    _validate_cli_args as _validate_cli_args_impl,
+)
+from src.cli import (
+    build_arg_parser as build_arg_parser_impl,
+)
+from src.cli import (
+    parse_arguments as parse_arguments_impl,
+)
+from src.cli import (
+    resolve_article_path as resolve_article_path_impl,
+)
+from src.cli import (
+    resolve_output_paths as resolve_output_paths_impl,
+)
+from src.cli import (
+    validate_imagedir as validate_imagedir_impl,
+)
 
 # Import config FIRST to set telemetry/system env vars before any library imports
 from src.config import config, validate_environment_variables
@@ -37,8 +59,55 @@ from src.eval import (
     reset_active_capture_manager,
     set_active_capture_manager,
 )
-from src.report_generator import QuietModeReporter
-from src.runtime_diagnostics import build_analysis_validity, is_publishable_analysis
+from src.output import (
+    _emit_start_banner as _emit_start_banner_impl,
+)
+from src.output import (
+    _load_company_name_for_output as _load_company_name_for_output_impl,
+)
+from src.output import (
+    _maybe_generate_article as _maybe_generate_article_impl,
+)
+from src.output import (
+    _render_primary_output as _render_primary_output_impl,
+)
+from src.output import (
+    _report_analysis_failure as _report_analysis_failure_impl,
+)
+from src.output import (
+    display_memory_statistics as display_memory_statistics_impl,
+)
+from src.output import (
+    display_results as display_results_impl,
+)
+from src.output import (
+    display_token_summary as display_token_summary_impl,
+)
+from src.output import (
+    display_welcome_banner as display_welcome_banner_impl,
+)
+from src.output import (
+    get_welcome_banner as get_welcome_banner_impl,
+)
+from src.output import (
+    handle_article_generation as handle_article_generation_impl,
+)
+from src.persistence import (
+    _maybe_save_rejection_record as _maybe_save_rejection_record_impl,
+)
+from src.persistence import (
+    _normalize_macro_context_metadata as _normalize_macro_context_metadata_impl,
+)
+from src.persistence import (
+    _persist_analysis_outputs as _persist_analysis_outputs_impl,
+)
+from src.persistence import (
+    build_run_summary as build_run_summary_impl,
+)
+from src.persistence import (
+    save_results_to_file as save_results_to_file_impl,
+)
+from src.runtime_diagnostics import build_analysis_validity
 
 # IMPORTANT: Don't import get_tracker here - it instantiates the singleton immediately
 # Import it lazily in functions that need it, after quiet mode is set
@@ -60,19 +129,6 @@ CLI_NOISY_DEPENDENCY_LOGGERS: dict[str, int] = {
     "urllib3": logging.WARNING,
 }
 HTTP_TRACE_LOGGERS = ("openai", "httpx", "httpcore", "hpack")
-
-
-@dataclass(frozen=True)
-class OutputTargets:
-    """Resolved output and chart destinations for a CLI run."""
-
-    output_file: Path | None
-    image_dir: Path
-    skip_charts: bool
-
-    @property
-    def output_dir(self) -> Path:
-        return self.output_file.parent if self.output_file else Path.cwd()
 
 
 def _cost_suffix() -> str:
@@ -137,326 +193,27 @@ def suppress_all_logging():
 def _cli_logging_mode(
     args,
 ) -> Literal["quiet", "brief", "normal", "verbose", "debug"]:
-    if getattr(args, "quiet", False):
-        return "quiet"
-    if getattr(args, "brief", False):
-        return "brief"
-    if getattr(args, "debug", False):
-        return "debug"
-    if getattr(args, "verbose", False):
-        return "verbose"
-    return "normal"
+    return _cli_logging_mode_impl(args)
 
 
 def build_arg_parser() -> argparse.ArgumentParser:
-    """Build and return the argument parser (separated for testability)."""
-    parser = argparse.ArgumentParser(
-        description="Multi-Agent Investment Analysis System (Gemini 3 Edition)",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  # Basic analysis
-  poetry run python -m src.main --ticker AAPL
-
-  # Quick analysis mode (Gemini Flash)
-  poetry run python -m src.main --ticker NVDA --quick
-
-  # Strict quality gate (tighter thresholds, fewer BUYs, token savings on rejects)
-  poetry run python -m src.main --ticker 0005.HK --strict
-
-  # Composable: strict quality bar + quick/cheap models
-  poetry run python -m src.main --ticker 0005.HK --strict --quick
-
-  # Quiet mode (markdown report only)
-  poetry run python -m src.main --ticker AAPL --quiet
-
-  # Brief mode (header, summary, decision only)
-  poetry run python -m src.main --ticker AAPL --brief
-
-  # Custom models
-  poetry run python -m src.main --ticker TSLA --quick-model gemini-2.5-flash --deep-model gemini-3-pro-preview
-
-  # Enable Langfuse tracing for this run
-  poetry run python -m src.main --ticker 0005.HK --enable-langfuse
-
-  # Batch retrospective: process all past tickers
-  poetry run python -m src.main --retrospective-only
-
-  # Activated venv alternative
-  python -m src.main --ticker MSFT --quick
-        """,
-    )
-
-    parser.add_argument(
-        "--ticker",
-        type=str,
-        required=False,
-        default=None,
-        help="Stock ticker symbol to analyze (e.g., AAPL, NVDA, TSLA)",
-    )
-
-    parser.add_argument(
-        "--quick",
-        action="store_true",
-        help="Use quick analysis mode (faster, less detailed)",
-    )
-
-    parser.add_argument(
-        "--strict",
-        action="store_true",
-        default=False,
-        help=(
-            "Apply stricter financial health criteria: tighter D/E and coverage thresholds, "
-            "auto-reject REITs/ETFs, PFIC, and VIE structures, escalate value-trap warnings "
-            "to rejects, and require higher conviction for BUY. Reduces BUY count and saves "
-            "tokens by rejecting candidates before Bull/Bear debate. Composable with --quick."
-        ),
-    )
-
-    parser.add_argument(
-        "--quiet",
-        action="store_true",
-        help="Suppress all logging and output only final markdown report",
-    )
-
-    parser.add_argument(
-        "--brief",
-        action="store_true",
-        help="Output only header, executive summary, and decision rationale",
-    )
-
-    parser.add_argument(
-        "--quick-model",
-        type=str,
-        default=None,
-        help=f"Model to use for quick analysis (default: {config.quick_think_llm})",
-    )
-
-    parser.add_argument(
-        "--deep-model",
-        type=str,
-        default=None,
-        help=f"Model to use for deep analysis (default: {config.deep_think_llm})",
-    )
-
-    parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Enable high-signal application diagnostics",
-    )
-
-    parser.add_argument(
-        "--debug",
-        action="store_true",
-        help=(
-            "Enable developer debug logging; use INVESTMENT_AGENT_TRACE_HTTP=1 "
-            "for raw transport traces"
-        ),
-    )
-
-    parser.add_argument(
-        "--no-memory", action="store_true", help="Disable persistent memory (ChromaDB)"
-    )
-
-    parser.add_argument(
-        "--svg",
-        action="store_true",
-        help="Generate charts in SVG format (default: PNG)",
-    )
-
-    parser.add_argument(
-        "--transparent",
-        action="store_true",
-        help="Use transparent background for charts (default: white grid)",
-    )
-
-    parser.add_argument(
-        "--no-charts",
-        action="store_true",
-        help="Skip chart generation entirely",
-    )
-
-    parser.add_argument(
-        "--enable-langfuse",
-        action="store_true",
-        help=(
-            "Enable Langfuse tracing for this run. Requires LANGFUSE_PUBLIC_KEY "
-            "and LANGFUSE_SECRET_KEY."
-        ),
-    )
-
-    parser.add_argument(
-        "--trace-langfuse",
-        action="store_true",
-        help=argparse.SUPPRESS,
-    )
-
-    parser.add_argument(
-        "--output",
-        type=str,
-        default=None,
-        help="Output file path (default: stdout). If set, images will default to {output_dir}/images",
-    )
-
-    parser.add_argument(
-        "--imagedir",
-        type=str,
-        default=None,
-        help="Directory for chart images. If --output is set, defaults to {output_dir}/images. If not set, defaults to 'images' in current dir.",
-    )
-
-    parser.add_argument(
-        "--article",
-        nargs="?",
-        const=True,
-        default=False,
-        help=(
-            "Generate a Medium-style article from the analysis. "
-            "Can specify output path (e.g., --article article.md) or use default "
-            "(e.g., --article generates {ticker}_article.md in results dir)."
-        ),
-    )
-
-    parser.add_argument(
-        "--retrospective-only",
-        action="store_true",
-        help="Run retrospective evaluation on all past analyses without running "
-        "a new analysis. Processes all tickers found in results directory.",
-    )
-
-    parser.add_argument(
-        "--capture-baseline",
-        action="store_true",
-        help=(
-            "Capture a versioned baseline bundle for this run under evals/captures/. "
-            "Does not perform evaluation or baseline promotion."
-        ),
-    )
-
-    parser.add_argument(
-        "--capture-baseline-cleanup",
-        action="store_true",
-        help=(
-            "Clean up stale inflight baseline captures under evals/captures/ and exit, "
-            "or run cleanup before capture when combined with --capture-baseline."
-        ),
-    )
-
-    return parser
+    return build_arg_parser_impl()
 
 
 def parse_arguments() -> argparse.Namespace:
-    """Parse command line arguments."""
-    parser = build_arg_parser()
-    args = parser.parse_args()
-
-    # Validate: --ticker is required unless --retrospective-only or cleanup-only
-    if (
-        not args.retrospective_only
-        and not args.capture_baseline_cleanup
-        and not args.ticker
-    ):
-        parser.error(
-            "--ticker is required unless --retrospective-only or "
-            "--capture-baseline-cleanup is specified"
-        )
-
-    if args.debug:
-        args.verbose = True
-
-    return args
+    return parse_arguments_impl()
 
 
 def resolve_output_paths(args) -> tuple[Path | None, Path]:
-    """
-    Determine output file and image directory based on arguments.
-
-    Args:
-        args: Parsed arguments namespace
-
-    Returns:
-        Tuple of (output_file_path, image_dir_path)
-    """
-    # Determine output location
-    output_file = Path(args.output) if args.output else None
-    output_dir = output_file.parent if output_file else Path.cwd()
-
-    # Determine image directory
-    if args.imagedir:
-        # User specified image directory
-        image_dir = Path(args.imagedir)
-    elif output_file:
-        # Default to {output_dir}/images if writing to file
-        image_dir = output_dir / "images"
-    else:
-        # Default to current directory's images if stdout
-        image_dir = Path("images")
-
-    return output_file, image_dir
+    return resolve_output_paths_impl(args)
 
 
 def validate_imagedir(imagedir: str) -> Path:
-    """Validate image directory path.
-
-    Allow any path (relative or absolute).
-    """
-    return Path(imagedir)
+    return validate_imagedir_impl(imagedir)
 
 
 def resolve_article_path(args, ticker: str) -> Path | None:
-    """
-    Determine article output path based on arguments.
-
-    Args:
-        args: Parsed arguments namespace
-        ticker: Stock ticker symbol
-
-    Returns:
-        Path for article output, or None if --article not specified
-
-    Path resolution logic:
-        1. --article /abs/path.md  -> Use absolute path as-is
-        2. --article rel.md --output /dir/report.md  -> /dir/rel.md (relative to output dir)
-        3. --article rel.md (no --output)  -> rel.md (relative to cwd)
-        4. --article --output /dir/report.md  -> /dir/report-ARTICLE.md
-        5. --article (no --output)  -> results/{ticker}_article.md
-    """
-    if not args.article:
-        return None
-
-    if isinstance(args.article, str):
-        # --article with explicit path
-        article_path = Path(args.article)
-        if not article_path.suffix:
-            article_path = article_path.with_suffix(".md")
-
-        # If absolute path, use as-is
-        if article_path.is_absolute():
-            return article_path
-
-        # If relative path and --output specified, resolve relative to output directory
-        if args.output:
-            output_dir = Path(args.output).parent
-            return output_dir / article_path
-
-        # Otherwise, relative to current working directory
-        return article_path
-
-    elif args.article is True:
-        # --article with no value
-        if args.output:
-            # Derive from --output path: add "_article" suffix (consistent with image naming)
-            output_path = Path(args.output)
-            stem = output_path.stem  # e.g., "0005_HK_2026-01-01"
-            suffix = output_path.suffix or ".md"  # e.g., ".md"
-            article_name = f"{stem}_article{suffix}"
-            return output_path.parent / article_name
-        else:
-            # No --output: use default path in results dir
-            safe_ticker = ticker.replace(".", "_").replace("/", "_")
-            return config.results_dir / f"{safe_ticker}_article.md"
-    else:
-        return None
+    return resolve_article_path_impl(args, ticker)
 
 
 def run_provider_preflight() -> dict[str, dict[str, str]]:
@@ -591,107 +348,12 @@ def build_run_summary(
     article_requested: bool,
     provider_preflight: dict[str, dict[str, str]] | None = None,
 ) -> dict[str, object]:
-    """Build a compact summary for saved artifacts and end-of-run logs."""
-    from langchain_core.messages import ToolMessage
-
-    from src.token_tracker import get_tracker
-
-    def _tool_message_failed(content: object) -> bool:
-        if not isinstance(content, str):
-            return False
-        text = content.strip()
-        if not text:
-            return False
-        if text.startswith(
-            (
-                "TOOL_ERROR:",
-                "TOOL_BLOCKED:",
-                "FETCH_FAILED:",
-                "SEARCH_FAILED:",
-                "INVALID_URL:",
-            )
-        ):
-            return True
-        try:
-            payload = json.loads(text)
-        except (TypeError, ValueError):
-            return False
-        return isinstance(payload, dict) and bool(payload.get("error"))
-
-    def _collect_used_providers() -> list[str]:
-        providers: set[str] = set()
-        configured = str(config.llm_provider or "").strip()
-        if configured:
-            providers.add(configured)
-        artifact_statuses = result.get("artifact_statuses", {}) or {}
-        for status in artifact_statuses.values():
-            provider = str((status or {}).get("provider") or "").strip()
-            if provider:
-                providers.add(provider)
-        return sorted(providers)
-
-    manual_tool_failures = sum(
-        value
-        for key, value in result.items()
-        if key.endswith("_tool_failures") and isinstance(value, int) and value > 0
+    return build_run_summary_impl(
+        result,
+        quick_mode=quick_mode,
+        article_requested=article_requested,
+        provider_preflight=provider_preflight,
     )
-
-    tracker_stats = get_tracker().get_total_stats()
-    messages = result.get("messages", []) or []
-    tool_messages = [msg for msg in messages if isinstance(msg, ToolMessage)]
-    tool_failures = manual_tool_failures + sum(
-        1
-        for msg in tool_messages
-        if getattr(msg, "status", None) == "error" or _tool_message_failed(msg.content)
-    )
-    artifact_statuses = result.get("artifact_statuses", {}) or {}
-    consultant_status = artifact_statuses.get("consultant_review") or {}
-    auditor_status = artifact_statuses.get("auditor_report") or {}
-    consultant_finished = bool(consultant_status.get("complete"))
-    auditor_finished = bool(auditor_status.get("complete"))
-    providers_used = _collect_used_providers()
-
-    return {
-        "quick_mode": quick_mode,
-        "quick_model": config.quick_think_llm,
-        "deep_model": config.deep_think_llm,
-        "provider_preflight": provider_preflight or {},
-        "pre_screening_result": result.get("pre_screening_result", ""),
-        "debate_rounds": result.get("investment_debate_state", {}).get("count", 0),
-        # Backward-compatible aliases: "completed" here means "finished", not "succeeded".
-        "consultant_completed": consultant_finished,
-        "auditor_completed": auditor_finished,
-        "consultant_finished": consultant_finished,
-        "auditor_finished": auditor_finished,
-        "consultant_successful": bool(consultant_status.get("ok")),
-        "auditor_successful": bool(auditor_status.get("ok")),
-        "article_requested": article_requested,
-        "llm_attempts": tracker_stats["total_calls"] + tracker_stats["failed_attempts"],
-        "llm_failures": tracker_stats["failed_attempts"],
-        "tool_calls": len(tool_messages),
-        "tool_failures": tool_failures,
-        "llm_providers_used": providers_used,
-        "llm_provider": providers_used[0]
-        if len(providers_used) == 1
-        else "multi-provider",
-        "macro_context_status": result.get("macro_context_status", "failed"),
-        "macro_context_region": result.get("macro_context_region", "GLOBAL"),
-        "macro_context_report_present": bool(result.get("macro_context_report")),
-        "macro_context_injected_into_news": bool(
-            result.get("macro_context_injected_into_news", False)
-        ),
-        "publishable": result.get("analysis_validity", {}).get("publishable", False),
-        "required_failures": sorted(
-            (result.get("analysis_validity", {}) or {})
-            .get("required_failures", {})
-            .keys()
-        ),
-        "optional_failures": sorted(
-            (result.get("analysis_validity", {}) or {})
-            .get("optional_failures", {})
-            .keys()
-        ),
-    }
 
 
 def _normalize_macro_context_metadata(
@@ -699,38 +361,7 @@ def _normalize_macro_context_metadata(
     *,
     cache_dir: Path | str | None = None,
 ) -> dict[str, Any]:
-    """Return one normalized macro-context metadata view for persistence.
-
-    ``cache_dir`` lets the saved artifact reflect the actual results destination
-    used for this run rather than ambient global config.
-    """
-    run_summary = result.get("run_summary", {}) or {}
-    status = result.get("macro_context_status")
-    region = result.get("macro_context_region")
-    report_present = result.get("macro_context_report")
-
-    if status is None:
-        status = run_summary.get("macro_context_status", "failed")
-    if region is None:
-        region = run_summary.get("macro_context_region", "GLOBAL")
-    if report_present is None:
-        report_present = run_summary.get("macro_context_report_present", False)
-
-    return {
-        "status": status,
-        "region": region,
-        "report_present": bool(report_present),
-        "injected_into_news": bool(
-            result.get("macro_context_injected_into_news", False)
-        ),
-        "llm_invoked": bool(result.get("macro_context_llm_invoked", False)),
-        "generated_at": result.get("macro_context_generated_at"),
-        "cache_dir": str(
-            Path(cache_dir)
-            if cache_dir is not None
-            else Path(config.results_dir) / ".macro_context_cache"
-        ),
-    }
+    return _normalize_macro_context_metadata_impl(result, cache_dir=cache_dir)
 
 
 async def handle_article_generation(
@@ -744,322 +375,41 @@ async def handle_article_generation(
     tracing_callbacks: list[Any] | None = None,
     tracing_metadata: dict[str, Any] | None = None,
 ) -> None:
-    """
-    Generate article if --article flag is set, then run Editor-in-Chief review.
-
-    Args:
-        args: Parsed arguments namespace
-        ticker: Stock ticker symbol
-        company_name: Full company name
-        report_text: The full analysis report
-        trade_date: Date of the analysis
-        valuation_context: Optional context about chart valuation vs decision
-        analysis_result: Raw result dictionary containing DATA_BLOCK/PM_BLOCK
-    """
-    article_path = resolve_article_path(args, ticker)
-    if not article_path:
-        return
-
-    try:
-        from src.article_writer import ArticleEditor, ArticleWriter
-
-        if not args.quiet and not args.brief:
-            console.print("\n[cyan]Generating article...[/cyan]")
-
-        # Default to local paths so markdown renders immediately in editors
-        # Users who want GitHub URLs can set GITHUB_RAW_BASE env var
-        writer = ArticleWriter(
-            use_github_urls=False,
-            callbacks=tracing_callbacks,
-            tracing_metadata=tracing_metadata,
-        )
-        draft_article = writer.write(
-            ticker=ticker,
-            company_name=company_name,
-            report_text=report_text,
-            trade_date=trade_date,
-            output_path=article_path,
-            valuation_context=valuation_context,
-        )
-
-        # Run Editor-in-Chief review loop
-        editor = ArticleEditor(
-            callbacks=tracing_callbacks,
-            tracing_metadata=tracing_metadata,
-        )
-        final_article = draft_article  # Default to draft if editor unavailable or fails
-
-        if editor.is_available():
-            if not args.quiet and not args.brief:
-                console.print("[cyan]Running Editor-in-Chief review...[/cyan]")
-
-            # Extract ground truth from analysis result
-            data_block = ""
-            pm_block = ""
-            valuation_params = ""
-            if analysis_result:
-                data_block = analysis_result.get("fundamentals_report", "")
-                pm_block = analysis_result.get("final_trade_decision", "")
-                valuation_params = analysis_result.get("valuation_params", "")
-
-            try:
-                final_article, feedback = await editor.edit(
-                    writer=writer,
-                    article_draft=draft_article,
-                    ticker=ticker,
-                    company_name=company_name,
-                    data_block=data_block,
-                    pm_block=pm_block,
-                    valuation_params=valuation_params,
-                )
-
-                # Log editor outcome
-                if feedback.get("skipped"):
-                    logger.info("Editor skipped (not available)")
-                elif feedback.get("verdict") == "APPROVED":
-                    logger.info(
-                        "Article approved by editor",
-                        confidence=feedback.get("confidence"),
-                    )
-                else:
-                    logger.info(
-                        "Article revised by editor",
-                        revisions=feedback.get("revisions", 0),
-                    )
-
-                # Save the final (possibly edited) article
-                if final_article != draft_article:
-                    with open(article_path, "w") as f:
-                        f.write(final_article)
-                    if not args.quiet and not args.brief:
-                        console.print("[green]Article revised and saved.[/green]")
-
-            except Exception as e:
-                # Safety net: if editor fails, preserve the draft
-                logger.warning(
-                    f"Editor revision failed, preserving original draft: {e}"
-                )
-                final_article = draft_article
-                if not args.quiet and not args.brief:
-                    console.print(
-                        "[yellow]Editor revision failed, using original draft.[/yellow]"
-                    )
-
-        if not args.quiet and not args.brief:
-            console.print(
-                f"[green]Article saved to:[/green] [cyan]{article_path}[/cyan]{_cost_suffix()}"
-            )
-            # Defensive: ensure article is a string before counting words
-            word_count = (
-                len(final_article.split()) if isinstance(final_article, str) else 0
-            )
-            console.print(f"[dim]Word count: {word_count} words[/dim]")
-
-    except Exception as e:
-        logger.error(
-            "article_generation_failed",
-            **summarize_exception(
-                e,
-                operation="generating article",
-                provider="unknown",
-            ),
-            exc_info=True,
-        )
-        if not args.quiet and not args.brief:
-            console.print(
-                f"[yellow]Warning: {_safe_cli_error_message('generating article', e)}[/yellow]"
-            )
+    await handle_article_generation_impl(
+        args=args,
+        ticker=ticker,
+        company_name=company_name,
+        report_text=report_text,
+        trade_date=trade_date,
+        valuation_context=valuation_context,
+        analysis_result=analysis_result,
+        tracing_callbacks=tracing_callbacks,
+        tracing_metadata=tracing_metadata,
+        logger_obj=logger,
+        console_obj=console,
+        resolve_article_path_fn=resolve_article_path,
+        error_message_formatter=_safe_cli_error_message,
+    )
 
 
 def get_welcome_banner(ticker: str, quick_mode: bool) -> str:
-    """Generate welcome banner string with configuration."""
-    banner = []
-    banner.append("# Multi-Agent Investment Analysis System")
-    banner.append("")
-    banner.append(f"**Ticker:** {ticker.upper()}  ")
-    banner.append(f"**Analysis Mode:** {'Quick' if quick_mode else 'Deep'}  ")
-    banner.append(f"**Quick Model:** {config.quick_think_llm}  ")
-    banner.append(f"**Deep Model:** {config.deep_think_llm}  ")
-    banner.append(
-        f"**Memory System:** {'Enabled' if config.enable_memory else 'Disabled'}  "
-    )
-    banner.append(
-        f"**LangSmith Tracing:** "
-        f"{'Enabled' if config.langsmith_tracing_enabled else 'Disabled'}  "
-    )
-    banner.append(
-        f"**Langfuse Tracing:** "
-        f"{'Enabled' if config.langfuse_enabled else 'Disabled'}  "
-    )
-    banner.append("")
-    return "\n".join(banner)
+    return get_welcome_banner_impl(ticker, quick_mode)
 
 
 def display_welcome_banner(ticker: str, quick_mode: bool):
-    """Display welcome banner with configuration.
-
-    Deprecated: Use get_welcome_banner() instead.
-    """
-    print(get_welcome_banner(ticker, quick_mode))
+    display_welcome_banner_impl(ticker, quick_mode)
 
 
 def display_memory_statistics(ticker: str):
-    """Display memory statistics for the current ticker."""
-    if not config.enable_memory:
-        return
-
-    try:
-        from src.memory import create_memory_instances, sanitize_ticker_for_collection
-
-        # Get memories specific to THIS ticker
-        memories = create_memory_instances(ticker)
-        safe_ticker = sanitize_ticker_for_collection(ticker)
-
-        console.print(f"\n[bold cyan]Memory Statistics for {ticker}:[/bold cyan]\n")
-
-        memory_table = Table(show_header=True, box=box.ROUNDED)
-        memory_table.add_column("Agent", style="cyan")
-        memory_table.add_column("Available", style="yellow")
-        memory_table.add_column("Total Memories", style="green")
-        memory_table.add_column("Status", style="blue")
-
-        agent_mapping = [
-            ("Bull Researcher", f"{safe_ticker}_bull_memory"),
-            ("Bear Researcher", f"{safe_ticker}_bear_memory"),
-            ("Research Manager", f"{safe_ticker}_invest_judge_memory"),
-            ("Trader", f"{safe_ticker}_trader_memory"),
-            ("Portfolio Manager", f"{safe_ticker}_risk_manager_memory"),
-        ]
-
-        for display_name, mem_key in agent_mapping:
-            mem = memories.get(mem_key)
-            if mem:
-                stats = mem.get_stats()
-                available = "✓" if stats.get("available") else "✗"
-                total = str(stats.get("count", 0))
-                status = "Active" if stats.get("available") else "Inactive"
-                memory_table.add_row(display_name, available, total, status)
-
-        console.print(memory_table)
-        console.print()
-
-    except Exception as e:
-        logger.warning(
-            "memory_statistics_unavailable",
-            **summarize_exception(e, operation="display memory statistics"),
-        )
+    display_memory_statistics_impl(ticker, console_obj=console, logger_obj=logger)
 
 
 def display_token_summary():
-    """Display token usage summary in a formatted table."""
-    from src.token_tracker import get_tracker
-
-    tracker = get_tracker()
-    stats = tracker.get_total_stats()
-
-    if stats["total_calls"] == 0:
-        return
-
-    console.print("\n[bold cyan]Token Usage Summary:[/bold cyan]\n")
-
-    # Overall stats table
-    summary_table = Table(show_header=True, box=box.ROUNDED)
-    summary_table.add_column("Metric", style="cyan")
-    summary_table.add_column("Value", style="green", justify="right")
-
-    summary_table.add_row("Total LLM Calls", str(stats["total_calls"]))
-    summary_table.add_row("Total Prompt Tokens", f"{stats['total_prompt_tokens']:,}")
-    summary_table.add_row(
-        "Total Completion Tokens", f"{stats['total_completion_tokens']:,}"
-    )
-    summary_table.add_row("Total Tokens", f"{stats['total_tokens']:,}")
-    summary_table.add_row(
-        "Projected Cost (Paid Tier)", f"${stats['total_cost_usd']:.4f}"
-    )
-
-    console.print(summary_table)
-
-    # Per-agent breakdown
-    console.print("\n[bold cyan]Per-Agent Token Usage:[/bold cyan]\n")
-
-    agent_table = Table(show_header=True, box=box.ROUNDED)
-    agent_table.add_column("Agent", style="cyan")
-    agent_table.add_column("Calls", style="yellow", justify="right")
-    agent_table.add_column("Prompt Tokens", style="blue", justify="right")
-    agent_table.add_column("Completion Tokens", style="magenta", justify="right")
-    agent_table.add_column("Total Tokens", style="green", justify="right")
-    agent_table.add_column("Cost (USD)", style="red", justify="right")
-
-    # Sort by cost descending
-    sorted_agents = sorted(
-        stats["agents"].items(), key=lambda x: x[1]["cost_usd"], reverse=True
-    )
-
-    for agent_name, agent_stats in sorted_agents:
-        agent_table.add_row(
-            agent_name,
-            str(agent_stats["calls"]),
-            f"{agent_stats['prompt_tokens']:,}",
-            f"{agent_stats['completion_tokens']:,}",
-            f"{agent_stats['total_tokens']:,}",
-            f"${agent_stats['cost_usd']:.4f}",
-        )
-
-    console.print(agent_table)
-    console.print()
+    display_token_summary_impl(console_obj=console)
 
 
 def display_results(result: dict, ticker: str):
-    """Display analysis results in a formatted manner."""
-    console.print("\n" + "=" * 80)
-    console.print("[bold green]Analysis Complete![/bold green]\n")
-
-    # Display token usage first
-    display_token_summary()
-
-    # Display final trading decision
-    if "final_trade_decision" in result and result["final_trade_decision"]:
-        decision_panel = Panel(
-            result["final_trade_decision"],
-            title="Final Trading Decision",
-            border_style="green",
-            padding=(1, 2),
-        )
-        console.print(decision_panel)
-
-    # Display individual analyst reports
-    console.print("\n[bold cyan]Analyst Reports:[/bold cyan]\n")
-
-    report_fields = [
-        ("market_report", "Market Analysis"),
-        ("sentiment_report", "Sentiment Analysis"),
-        ("news_report", "News Analysis"),
-        ("foreign_language_report", "Foreign Language Analysis"),
-        ("fundamentals_report", "Fundamentals Analysis"),
-        ("investment_plan", "Investment Plan"),
-        ("trader_investment_plan", "Trading Proposal"),
-    ]
-
-    for field_name, display_name in report_fields:
-        if field_name in result and result[field_name]:
-            content = result[field_name]
-
-            if content.startswith("Error"):
-                style = "red"
-            else:
-                style = "cyan"
-
-            if len(content) > 800:
-                content = content[:800] + "\n\n[... truncated for display ...]"
-
-            report_panel = Panel(
-                content, title=f"{display_name}", border_style=style, padding=(1, 2)
-            )
-            console.print(report_panel)
-            console.print()
-
-    display_memory_statistics(ticker)
-    console.print("=" * 80 + "\n")
+    display_results_impl(result, ticker, console_obj=console)
 
 
 def save_results_to_file(
@@ -1070,270 +420,14 @@ def save_results_to_file(
     results_dir: Path | str | None = None,
     trace_id: str | None = None,
 ) -> Path:
-    """Save analysis results to a JSON file in the results directory."""
-    from src.memory import get_ticker_memory_stats
-    from src.prompts import get_all_prompts
-
-    results_dir = (
-        Path(results_dir) if results_dir is not None else Path(config.results_dir)
+    return save_results_to_file_impl(
+        result,
+        ticker,
+        quick_mode=quick_mode,
+        results_dir=results_dir,
+        trace_id=trace_id,
+        logger_obj=logger,
     )
-    results_dir.mkdir(parents=True, exist_ok=True)
-    previous_dir_mtime_ns = (
-        results_dir.stat().st_mtime_ns if results_dir.exists() else None
-    )
-    analysis_file_count_before_save = sum(
-        1 for candidate in results_dir.glob("*_analysis.json") if candidate.is_file()
-    )
-
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    filename = f"{ticker}_{timestamp}_analysis.json"
-    filepath = results_dir / filename
-
-    prompts_used = result.get("prompts_used", {})
-    all_prompts = get_all_prompts()
-    available_prompts = {
-        key: {
-            "agent_name": prompt.agent_name,
-            "version": prompt.version,
-            "category": prompt.category,
-            "requires_tools": prompt.requires_tools,
-        }
-        for key, prompt in all_prompts.items()
-    }
-
-    prompts_dir = Path("./prompts")
-    custom_prompts_loaded = []
-    if prompts_dir.exists():
-        for json_file in prompts_dir.glob("*.json"):
-            custom_prompts_loaded.append(json_file.stem)
-
-    memory_stats = {}
-    if config.enable_memory:
-        try:
-            memory_stats = get_ticker_memory_stats(ticker)
-        except Exception as e:
-            logger.warning(
-                "memory_stats_unavailable",
-                **summarize_exception(e, operation="save memory stats"),
-            )
-
-    # Get token usage stats
-    from src.token_tracker import get_tracker
-
-    tracker = get_tracker()
-    token_stats = tracker.get_total_stats()
-
-    save_data = {
-        "metadata": {
-            "ticker": ticker,
-            "timestamp": timestamp,
-            "analysis_date": datetime.now().isoformat(),
-            "environment": config.environment,
-            "quick_model": config.quick_think_llm,
-            "deep_model": config.deep_think_llm,
-            "memory_enabled": config.enable_memory,
-            "online_tools_enabled": config.online_tools,
-            "llm_provider": (
-                (result.get("run_summary", {}) or {}).get("llm_provider")
-                or config.llm_provider
-            ),
-            "llm_providers_used": (
-                (result.get("run_summary", {}) or {}).get("llm_providers_used")
-                or [config.llm_provider]
-            ),
-        },
-        "token_usage": token_stats,
-        "macro_context": _normalize_macro_context_metadata(
-            result,
-            cache_dir=results_dir / ".macro_context_cache",
-        ),
-        "prompts_metadata": {
-            "prompts_used": prompts_used,
-            "available_prompts": available_prompts,
-            "custom_prompts_loaded": custom_prompts_loaded,
-            "prompts_directory": str(prompts_dir),
-            "total_agents": len(prompts_used),
-            "note": (
-                "system_message field contains the actual prompt text used by each "
-                "graph agent or pre-graph helper"
-            ),
-        },
-        "memory_statistics": memory_stats,
-        "reports": {
-            "market_report": result.get("market_report", ""),
-            "sentiment_report": result.get("sentiment_report", ""),
-            "news_report": result.get("news_report", ""),
-            "fundamentals_report": result.get("fundamentals_report", ""),
-        },
-        "investment_analysis": {
-            "investment_debate": {
-                "bull_history": result.get("investment_debate_state", {}).get(
-                    "bull_history", ""
-                ),
-                "bear_history": result.get("investment_debate_state", {}).get(
-                    "bear_history", ""
-                ),
-                "debate_rounds": result.get("investment_debate_state", {}).get(
-                    "count", 0
-                ),
-            },
-            "investment_plan": result.get("investment_plan", ""),
-            "trader_plan": result.get("trader_investment_plan", ""),
-        },
-        "risk_analysis": {
-            "risk_debate": {
-                "risky_perspective": result.get("risk_debate_state", {}).get(
-                    "current_risky_response", ""
-                ),
-                "safe_perspective": result.get("risk_debate_state", {}).get(
-                    "current_safe_response", ""
-                ),
-                "neutral_perspective": result.get("risk_debate_state", {}).get(
-                    "current_neutral_response", ""
-                ),
-                "debate_rounds": 1,  # Risk analysts run in parallel (1 round each)
-            }
-        },
-        "final_decision": {
-            "decision": result.get("final_trade_decision", ""),
-            "processed_signal": None,
-        },
-        "pre_screening_result": result.get("pre_screening_result", ""),
-        "run_summary": result.get("run_summary", {}),
-        "analysis_validity": result.get("analysis_validity", {}),
-        "artifact_statuses": result.get("artifact_statuses", {}),
-    }
-
-    run_summary = save_data.get("run_summary", {}) or {}
-    macro_context_payload = save_data.get("macro_context", {}) or {}
-    prompt_records = (save_data.get("prompts_metadata", {}) or {}).get(
-        "prompts_used", {}
-    ) or {}
-    token_agents = (save_data.get("token_usage", {}) or {}).get("agents", {}) or {}
-    has_run_summary_macro_fields = all(
-        field in run_summary
-        for field in (
-            "macro_context_status",
-            "macro_context_region",
-            "macro_context_report_present",
-            "macro_context_injected_into_news",
-        )
-    )
-    has_macro_context_block = all(
-        field in macro_context_payload
-        for field in (
-            "status",
-            "region",
-            "report_present",
-            "injected_into_news",
-            "llm_invoked",
-            "generated_at",
-            "cache_dir",
-        )
-    )
-    has_macro_prompt_metadata = "macro_context_analyst" in prompt_records
-    has_macro_token_row = "Macro Context Analyst" in token_agents
-
-    logger.info(
-        "analysis_artifact_macro_snapshot",
-        ticker=ticker,
-        has_macro_context_block=has_macro_context_block,
-        has_run_summary_macro_fields=has_run_summary_macro_fields,
-        has_macro_prompt_metadata=has_macro_prompt_metadata,
-        has_macro_token_row=has_macro_token_row,
-    )
-
-    macro_expected = bool(
-        result.get("macro_context_llm_invoked", False)
-        or result.get("macro_context_injected_into_news", False)
-        or result.get("macro_context_report")
-    )
-    macro_mismatch = macro_expected and (
-        not has_macro_context_block
-        or not has_run_summary_macro_fields
-        or (
-            result.get("macro_context_llm_invoked", False)
-            and not has_macro_prompt_metadata
-        )
-        or (result.get("macro_context_llm_invoked", False) and not has_macro_token_row)
-    )
-    if macro_mismatch:
-        logger.warning(
-            "analysis_artifact_macro_mismatch",
-            ticker=ticker,
-            macro_expected=macro_expected,
-            macro_llm_invoked=bool(result.get("macro_context_llm_invoked", False)),
-            macro_context_injected_into_news=bool(
-                result.get("macro_context_injected_into_news", False)
-            ),
-            has_macro_context_block=has_macro_context_block,
-            has_run_summary_macro_fields=has_run_summary_macro_fields,
-            has_macro_prompt_metadata=has_macro_prompt_metadata,
-            has_macro_token_row=has_macro_token_row,
-        )
-
-    # Extract prediction snapshot for future retrospective evaluation (zero LLM cost)
-    try:
-        from src.retrospective import extract_snapshot
-
-        save_data["prediction_snapshot"] = extract_snapshot(
-            result,
-            ticker,
-            quick_mode,
-            trace_id=trace_id,
-        )
-    except Exception as e:
-        logger.warning(
-            "snapshot_extraction_failed",
-            **summarize_exception(e, operation="prediction snapshot extraction"),
-        )
-
-    with open(filepath, "w") as f:
-        json.dump(save_data, f, indent=2)
-
-    try:
-        from src.ibkr.reconciler import (
-            _build_analysis_record_from_data,
-            load_latest_analyses,
-            update_latest_analyses_index,
-        )
-
-        record = _build_analysis_record_from_data(filepath, save_data)
-        if record is not None:
-            updated_index = update_latest_analyses_index(
-                results_dir,
-                record,
-                previous_dir_mtime_ns=previous_dir_mtime_ns,
-                analysis_file_count_before_save=analysis_file_count_before_save,
-            )
-            if not updated_index:
-                refreshed = load_latest_analyses(results_dir)
-                logger.info(
-                    "analysis_index_refreshed_after_save",
-                    ticker=ticker,
-                    path=str(results_dir),
-                    refreshed_count=len(refreshed),
-                )
-    except Exception as exc:
-        logger.debug(
-            "analysis_index_update_skipped",
-            **summarize_exception(exc, operation="analysis index update"),
-        )
-
-    logger.info(
-        f"Results saved to {filepath} ({len(prompts_used)} prompts tracked, {len(custom_prompts_loaded)} custom)"
-    )
-
-    # Log token tracking info
-    if token_stats["total_calls"] > 0:
-        logger.info(
-            f"Token usage tracked: {token_stats['total_calls']} LLM calls, "
-            f"{token_stats['total_tokens']:,} total tokens, "
-            f"${token_stats['total_cost_usd']:.4f} projected cost (paid tier) - "
-            f"saved to {filepath}"
-        )
-
-    return filepath
 
 
 _BENCH_NAMES: dict[str, str] = {
@@ -1764,39 +858,11 @@ def _apply_runtime_overrides(args: argparse.Namespace) -> None:
 
 
 def _validate_cli_args(args: argparse.Namespace) -> None:
-    """Validate incompatible flag combinations."""
-    if not args.quick:
-        return
-
-    chart_flags = [
-        f
-        for f, value in [("--transparent", args.transparent), ("--svg", args.svg)]
-        if value
-    ]
-    if not chart_flags:
-        return
-
-    flags_str = " and ".join(chart_flags)
-    verb = "has" if len(chart_flags) == 1 else "have"
-    noun = "that flag" if len(chart_flags) == 1 else "those flags"
-    print(
-        f"error: {flags_str} {verb} no effect with --quick "
-        f"(chart generation is skipped in quick mode). "
-        f"Remove {noun} or drop --quick.",
-        file=sys.stderr,
-    )
-    raise SystemExit(2)
+    _validate_cli_args_impl(args)
 
 
 def _resolve_output_targets(args: argparse.Namespace) -> OutputTargets:
-    """Resolve output/image paths and derived chart behavior for this run."""
-    output_file, image_dir = resolve_output_paths(args)
-    skip_charts = bool(args.no_charts or (output_file is None and not args.imagedir))
-    return OutputTargets(
-        output_file=output_file,
-        image_dir=image_dir,
-        skip_charts=skip_charts,
-    )
+    return _resolve_output_targets_impl(args)
 
 
 def _setup_runtime(
@@ -2024,17 +1090,13 @@ async def _maybe_run_ticker_retrospective(args: argparse.Namespace) -> None:
 
 
 def _emit_start_banner(args: argparse.Namespace, output_targets: OutputTargets) -> str:
-    """Render or log the startup banner and return it for file output."""
-    welcome_banner = get_welcome_banner(args.ticker, args.quick)
-    if not output_targets.output_file and not args.quiet and not args.brief:
-        print(welcome_banner)
-    if output_targets.output_file and not args.quiet and not args.brief:
-        logger.info(
-            "analysis_output_starting",
-            ticker=args.ticker,
-            output_path=str(output_targets.output_file),
-        )
-    return welcome_banner
+    return _emit_start_banner_impl(
+        args,
+        output_targets,
+        logger_obj=logger,
+        print_fn=print,
+        welcome_banner_fn=get_welcome_banner,
+    )
 
 
 async def _execute_analysis(
@@ -2230,34 +1292,10 @@ def _score_analysis_trace(result: dict, trace_context: Any) -> None:
 
 
 def _load_company_name_for_output(ticker: str) -> str | None:
-    """Best-effort company-name lookup for markdown output contexts."""
-    try:
-        import yfinance as yf
-
-        from src.ticker_utils import (
-            _company_name_lookup_candidates,
-            _is_valid_company_name,
-            normalize_company_name,
-        )
-
-        with ThreadPoolExecutor(max_workers=1) as executor:
-            for lookup_ticker, _lookup_strategy in _company_name_lookup_candidates(
-                ticker
-            ):
-                future = executor.submit(
-                    lambda symbol=lookup_ticker: yf.Ticker(symbol).info
-                )
-                info = future.result(timeout=5)
-                if not info:
-                    continue
-                raw_name = info.get("longName") or info.get("shortName")
-                if _is_valid_company_name(raw_name, lookup_ticker):
-                    return normalize_company_name(raw_name)
-        return None
-    except FuturesTimeoutError:
-        return None
-    except Exception:
-        return None
+    return _load_company_name_for_output_impl(
+        ticker,
+        thread_pool_executor_cls=ThreadPoolExecutor,
+    )
 
 
 def _render_primary_output(
@@ -2265,65 +1303,18 @@ def _render_primary_output(
     args: argparse.Namespace,
     output_targets: OutputTargets,
     welcome_banner: str,
-) -> tuple[str | None, str | None, QuietModeReporter | None]:
-    """Render the main user-facing report to stdout or file."""
-    use_markdown = (
-        args.brief
-        or args.quiet
-        or not sys.stdout.isatty()
-        or output_targets.output_file
+) -> tuple[str | None, str | None, Any | None]:
+    return _render_primary_output_impl(
+        result,
+        args,
+        output_targets,
+        welcome_banner,
+        console_obj=console,
+        logger_obj=logger,
+        company_name_loader=_load_company_name_for_output,
+        display_results_fn=display_results,
+        cost_suffix_fn=_cost_suffix,
     )
-    company_name = None
-    report = None
-    reporter = None
-
-    if not use_markdown:
-        display_results(result, args.ticker)
-        return company_name, report, reporter
-
-    company_name = _load_company_name_for_output(args.ticker)
-    reporter = QuietModeReporter(
-        args.ticker,
-        company_name,
-        quick_mode=args.quick,
-        chart_format="svg" if args.svg else "png",
-        transparent_charts=args.transparent,
-        skip_charts=output_targets.skip_charts,
-        image_dir=output_targets.image_dir,
-        report_dir=output_targets.output_dir,
-        report_stem=output_targets.output_file.stem
-        if output_targets.output_file
-        else None,
-    )
-    report = reporter.generate_report(result, brief_mode=args.brief)
-
-    if output_targets.output_file:
-        full_content = report
-        try:
-            if output_targets.output_file.parent != Path("."):
-                output_targets.output_file.parent.mkdir(parents=True, exist_ok=True)
-            with open(output_targets.output_file, "w") as f:
-                f.write(full_content)
-            if not args.quiet and not args.brief:
-                console.print(
-                    f"[green]Report saved to:[/green] [cyan]{output_targets.output_file}[/cyan]{_cost_suffix()}"
-                )
-        except Exception as exc:
-            logger.error(
-                "report_write_failed",
-                path=str(output_targets.output_file),
-                **summarize_exception(
-                    exc,
-                    operation="writing markdown report",
-                    provider="unknown",
-                ),
-                exc_info=True,
-            )
-            raise SystemExit(1) from exc
-    else:
-        print(report)
-
-    return company_name, report, reporter
 
 
 def _persist_analysis_outputs(
@@ -2332,33 +1323,15 @@ def _persist_analysis_outputs(
     *,
     trace_id: str | None = None,
 ) -> None:
-    """Persist JSON artifacts and rejection records."""
-    try:
-        filepath = save_results_to_file(
-            result,
-            args.ticker,
-            quick_mode=args.quick,
-            results_dir=Path(config.results_dir),
-            trace_id=trace_id,
-        )
-        if not args.quiet and not args.brief:
-            console.print(
-                f"[green]Results saved to:[/green] [cyan]{filepath}[/cyan]{_cost_suffix()}"
-            )
-    except Exception as exc:
-        logger.error(
-            "results_save_failed",
-            **summarize_exception(
-                exc,
-                operation="saving analysis results",
-                provider="unknown",
-            ),
-            exc_info=True,
-        )
-        if not args.quiet and not args.brief:
-            console.print(
-                f"\n[yellow]Warning: {_safe_cli_error_message('saving analysis results', exc)}[/yellow]\n"
-            )
+    _persist_analysis_outputs_impl(
+        result,
+        args,
+        trace_id=trace_id,
+        logger_obj=logger,
+        console_obj=console,
+        cost_suffix_fn=_cost_suffix,
+        error_message_formatter=_safe_cli_error_message,
+    )
 
 
 async def _maybe_save_rejection_record(
@@ -2367,33 +1340,12 @@ async def _maybe_save_rejection_record(
     *,
     trace_id: str | None = None,
 ) -> None:
-    """Persist non-BUY verdicts as retrospective rejection records."""
-    try:
-        from src.retrospective import (
-            create_lessons_memory,
-            extract_snapshot,
-            save_rejection_record,
-        )
-
-        snapshot = extract_snapshot(
-            result,
-            args.ticker,
-            is_quick_mode=args.quick,
-            trace_id=trace_id,
-        )
-        verdict = snapshot.get("verdict", "")
-        if verdict and verdict != "BUY":
-            rejection_memory = create_lessons_memory()
-            await save_rejection_record(snapshot, rejection_memory)
-    except Exception as exc:
-        logger.debug(
-            "rejection_record_save_skipped",
-            **summarize_exception(
-                exc,
-                operation="saving rejection record",
-                provider="unknown",
-            ),
-        )
+    await _maybe_save_rejection_record_impl(
+        result,
+        args,
+        trace_id=trace_id,
+        logger_obj=logger,
+    )
 
 
 async def _maybe_generate_article(
@@ -2402,68 +1354,24 @@ async def _maybe_generate_article(
     output_targets: OutputTargets,
     company_name: str | None,
     report: str | None,
-    reporter: QuietModeReporter | None,
+    reporter: Any | None,
     tracing_callbacks: list[Any] | None = None,
     tracing_metadata: dict[str, Any] | None = None,
 ) -> bool:
-    """Generate an article from a publishable analysis when requested."""
-    if not args.article:
-        return False
-
-    if not is_publishable_analysis(result):
-        logger.warning(
-            "article_generation_skipped_invalid_analysis",
-            ticker=args.ticker,
-            analysis_validity=result.get("analysis_validity", {}),
-        )
-        if not args.quiet and not args.brief:
-            console.print(
-                "[yellow]Skipping article generation because the analysis is incomplete or invalid.[/yellow]"
-            )
-        return False
-
-    if (
-        output_targets.skip_charts
-        and not output_targets.output_file
-        and not args.imagedir
-    ):
-        print(
-            "Warning: Article generated without images (stdout mode).",
-            file=sys.stderr,
-        )
-
-    trade_date = result.get("trade_date") or datetime.now().strftime("%Y-%m-%d")
-
-    if report is None or reporter is None:
-        if company_name is None:
-            company_name = _load_company_name_for_output(args.ticker) or args.ticker
-        reporter = QuietModeReporter(
-            args.ticker,
-            company_name,
-            quick_mode=args.quick,
-            chart_format="svg" if args.svg else "png",
-            transparent_charts=args.transparent,
-            skip_charts=output_targets.skip_charts,
-            image_dir=output_targets.image_dir,
-            report_dir=output_targets.output_dir,
-            report_stem=output_targets.output_file.stem
-            if output_targets.output_file
-            else None,
-        )
-        report = reporter.generate_report(result, brief_mode=False)
-
-    await handle_article_generation(
-        args=args,
-        ticker=args.ticker,
-        company_name=company_name or args.ticker,
-        report_text=report,
-        trade_date=trade_date,
-        valuation_context=reporter.get_valuation_context(),
-        analysis_result=result,
+    return await _maybe_generate_article_impl(
+        result,
+        args,
+        output_targets,
+        company_name,
+        report,
+        reporter,
         tracing_callbacks=tracing_callbacks,
         tracing_metadata=tracing_metadata,
+        logger_obj=logger,
+        console_obj=console,
+        company_name_loader=_load_company_name_for_output,
+        handle_article_generation_fn=handle_article_generation,
     )
-    return True
 
 
 def _log_final_summary(
@@ -2478,15 +1386,7 @@ def _log_final_summary(
 
 
 def _report_analysis_failure(args: argparse.Namespace) -> None:
-    """Print the standard top-level analysis failure message."""
-    if args.quiet or args.brief:
-        print(
-            "# Analysis Failed\n\nAn error occurred during analysis. Check logs for details."
-        )
-    else:
-        console.print(
-            "\n[bold red]Analysis failed. Check logs for details.[/bold red]\n"
-        )
+    _report_analysis_failure_impl(args, console_obj=console)
 
 
 async def main() -> int:
